update socket monitoring code

+++ src/util.c
@@ -57,7 +57,7 @@ const char* zmqBridge_generateSerial(unsigned long long* id)
    return strdup(temp);
 }
 
-const char* get_zmqEventName(int event)
+const char* get_zmqEventName(uint64_t event)
 {
    switch(event) {
       case ZMQ_EVENT_CONNECTED                  : return "CONNECTED";
@@ -80,7 +80,7 @@ const char* get_zmqEventName(int event)
 }
 
 // Note: this code must be synchronized w/get_zmqEventMask
-int get_zmqEventLogLevel(int event)
+int get_zmqEventLogLevel(uint64_t event)
 {
    switch(event) {
       case ZMQ_EVENT_CONNECTED                  : return MAMA_LOG_LEVEL_NORMAL;
@@ -106,9 +106,9 @@ int get_zmqEventLogLevel(int event)
 // Returns a bit-mask of events that are logged at specified log level.
 // Ensures that we don't ask for more events than we care about
 // Note: this code must be synchronized w/get_zmqEventLogLevel
-int get_zmqEventMask(int logLevel)
+uint64_t get_zmqEventMask(int logLevel)
 {
-   int eventMask = 0;
+   uint64_t eventMask = 0;
 
    switch(logLevel) {
